Add redhat ovalv2
