+++ main.go
@@ -26,6 +26,7 @@ import (
 	debianoval "github.com/aquasecurity/vuln-list-update/oval/debian"
 	oracleoval "github.com/aquasecurity/vuln-list-update/oval/oracle"
 	redhatoval "github.com/aquasecurity/vuln-list-update/oval/redhat"
+	redhatoval2 "github.com/aquasecurity/vuln-list-update/oval/redhat2"
 	"github.com/aquasecurity/vuln-list-update/photon"
 	"github.com/aquasecurity/vuln-list-update/redhat"
 	"github.com/aquasecurity/vuln-list-update/ubuntu"
@@ -39,7 +40,7 @@ const (
 )
 
 var (
-	target = flag.String("target", "", "update target (nvd, alpine, redhat, redhat-oval, debian, debian-oval, ubuntu, amazon, oracle-oval, suse-cvrf, photon, ghsa)")
+	target = flag.String("target", "", "update target (nvd, alpine, redhat, redhat-oval, debian, debian-oval, ubuntu, amazon, oracle-oval, suse-cvrf, photon, ghsa, redhat-oval2, cwe)")
 	years  = flag.String("years", "", "update years (only redhat)")
 )
 
@@ -167,6 +168,12 @@ func run() error {
 			return xerrors.Errorf("error in CWE update: %w", err)
 		}
 		commitMsg = "CWE Advisories"
+	case "redhat2-oval":
+		rh2 := redhatoval2.NewConfig()
+		if err := rh2.Update(); err != nil {
+			return xerrors.Errorf("error in RedHat Oval v2 update: %w", err)
+		}
+		commitMsg = "RedHat Oval v2"
 	default:
 		return xerrors.New("unknown target")
 	}
