+++ oval/redhat2/testdata/PULP_MANIFEST
@@ -1,0 +1,3 @@
+RHEL6/rhel-6-extras-including-unpatched.oval.xml.bz2,6a4e05e6c5ef90d23c4e1752ac5bf247e075d17b76c0f60f6aa02e68daa5a6a4,6347
+RHEL7/dotnet-3.1-including-unpatched.oval.xml.bz2,3afa7c45c0ccd21444a61f236c40a417386e09ce278cf85b049bc79d02d3c493,5817
+RHEL8/ansible-2-including-unpatched.oval.xml.bz2,d695814ee6ac7de65106ae26d7da85e6efe68ea5e951fefef5f2327393e3041c,5501
