+++ oval/redhat2/testdata/golden/ansible-2-including-unpatched/CVE-2020-10744.json
@@ -1,0 +1,92 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:202010744",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2020-10744 ansible: incomplete fix for CVE-2020-1733 (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2020-10744",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2020-10744"
+      }
+    ],
+    "Description": "DOCUMENTATION: An incomplete fix was found for the fix of the flaw CVE-2020-1733, Ansible: insecure temporary directory when running become_user from the become directive. The provided fix is insufficient to prevent the race condition on systems using ACLs and FUSE filesystems.\n            STATEMENT: Ansible Engine 2.7.18, 2.8.12, and 2.9.9 as well as previous versions are affected.\n\nAnsible Tower 3.4.5, 3.5.6 and 3.6.4 as well as previous versions are affected.\n\nRed Hat Gluster Storage 3 and Red Hat Ceph Storage 3 no longer maintain their own versions of Ansible. The fix will be provided from core Ansible. However, we still ship Ansible separately for Ceph Ubuntu.\n            MITIGATION: Currently, there is no mitigation for this issue.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:202010744006"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 8 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:202010744005"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744001"
+                  },
+                  {
+                    "Comment": "ansible is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible-test is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744003"
+                  },
+                  {
+                    "Comment": "ansible-test is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
