+++ oval/redhat2/testdata/golden/ansible-2-including-unpatched/CVE-2020-1734.json
@@ -1,0 +1,92 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:20201734",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2020-1734 ansible: shell enabled by default in a pipe lookup plugin subprocess (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2020-1734",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2020-1734"
+      }
+    ],
+    "Description": "DOCUMENTATION: A flaw was found in the pipe lookup plugin of ansible. Arbitrary commands can be run, when the pipe lookup plugin uses subprocess.Popen() with shell=True, by overwriting ansible facts and the variable is not escaped by quote plugin. An attacker could take advantage and run arbitrary commands by overwriting the ansible facts.\n            STATEMENT: Ansible Engine 2.7.16, 2.8.10, and 2.9.6 as well as previous versions are affected.\n\nAnsible Tower 3.4.5, 3.5.5 and 3.6.3 as well as previous versions are affected.\n            MITIGATION: This issue can be avoided by escaping variables which are used in the lookup.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:202010744006"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 8 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:202010744005"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744001"
+                  },
+                  {
+                    "Comment": "ansible is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible-test is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744003"
+                  },
+                  {
+                    "Comment": "ansible-test is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
