+++ oval/redhat2/testdata/golden/ansible-2-including-unpatched/CVE-2020-1738.json
@@ -1,0 +1,92 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:20201738",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2020-1738 ansible: module package can be selected by the ansible facts (low)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2020-1738",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2020-1738"
+      }
+    ],
+    "Description": "DOCUMENTATION: A flaw was found in Ansible Engine when the module package or service is used and the parameter 'use' is not specified. If a previous task is executed with a malicious user, the module sent can be selected by the attacker using the ansible facts file.\n            STATEMENT: Ansible Engine 2.7.16, 2.8.10, and 2.9.6 as well as previous versions are affected.\n\nAnsible Tower 3.4.5, 3.5.5 and 3.6.3 as well as previous versions are affected.\n            MITIGATION: Specify the parameter 'use' when possible on the package and service modules. Avoid using Ansible Collections on Ansible 2.8.9 or 2.7.16 (and any of the previous versions) as they are not rejecting python with no path (already fixed in 2.9.x).",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Low",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:202010744006"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 8 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:202010744005"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744001"
+                  },
+                  {
+                    "Comment": "ansible is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible-test is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744003"
+                  },
+                  {
+                    "Comment": "ansible-test is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
