+++ oval/redhat2/testdata/golden/ansible-2-including-unpatched/RHSA-2019-3927.json
@@ -1,0 +1,117 @@
+{
+  "Class": "patch",
+  "ID": "oval:com.redhat.rhsa:def:20193927",
+  "Version": "634",
+  "Metadata": {
+    "Title": "RHSA-2019:3927: ansible security and bug fix update (Moderate)",
+    "AffectedList": [
+      {
+        "Family": "unix",
+        "Platforms": [
+          "Red Hat Ansible Engine 2.9 for RHEL 8"
+        ]
+      }
+    ],
+    "References": [
+      {
+        "Source": "RHSA",
+        "RefID": "RHSA-2019:3927",
+        "RefURL": "https://access.redhat.com/errata/RHSA-2019:3927"
+      },
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2019-14864",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2019-14864"
+      }
+    ],
+    "Description": "Ansible is a simple model-driven configuration management, multi-node\ndeployment, and remote-task execution system. Ansible works over SSH and\ndoes not require any software or daemons to be installed on remote nodes.\nExtension modules can be written in any language and are transferred to\nmanaged machines automatically.\n\nSecurity Fix(es):\n\nAnsible: Splunk and Sumologic callback plugins leak sensitive data in logs\n(CVE-2019-14864)\n\nFor more details about the security issue(s), including the impact, a CVSS\nscore, acknowledgments, and other related information, refer to the CVE\npage(s) listed in the References section.\n\nThe following packages have been upgraded to a newer upstream version:\nansible (2.9.1)\n\nBug Fix(es):\n\nSee:\nhttps://github.com/ansible/ansible/blob/v2.9.1/changelogs/CHANGELOG-v2.9.rst\nfor details on bug fixes in this release.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "Copyright 2019 Red Hat, Inc.",
+      "Issued": {
+        "Date": "2019-11-20"
+      },
+      "Updated": {
+        "Date": "2019-11-20"
+      },
+      "Cves": [
+        {
+          "CveID": "CVE-2019-14864",
+          "Cvss2": "",
+          "Cvss3": "5.7/CVSS:3.0/AV:N/AC:L/PR:L/UI:R/S:U/C:H/I:N/A:N",
+          "Cwe": "CWE-117 CWE-532",
+          "Impact": "",
+          "Href": "https://access.redhat.com/security/cve/CVE-2019-14864",
+          "Public": "20191022"
+        }
+      ],
+      "Bugzilla": {
+        "Href": "https://bugzilla.redhat.com/1764148",
+        "ID": "1764148"
+      },
+      "AffectedCpeList": {
+        "Cpe": [
+          "cpe:/a:redhat:ansible_engine:2.9",
+          "cpe:/a:redhat:ansible_engine:2.9::el8"
+        ]
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:202010744006"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 8 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:202010744005"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible is earlier than 0:2.9.1-1.el8",
+                    "TestRef": "oval:com.redhat.rhsa:tst:20193927001"
+                  },
+                  {
+                    "Comment": "ansible is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible-test is earlier than 0:2.9.1-1.el8",
+                    "TestRef": "oval:com.redhat.rhsa:tst:20193927003"
+                  },
+                  {
+                    "Comment": "ansible-test is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
