+++ oval/redhat2/testdata/golden/ansible-2-including-unpatched/RHSA-2020-0215.json
@@ -1,0 +1,131 @@
+{
+  "Class": "patch",
+  "ID": "oval:com.redhat.rhsa:def:20200215",
+  "Version": "632",
+  "Metadata": {
+    "Title": "RHSA-2020:0215: Ansible security and bug fix update (2.9.4) (Moderate)",
+    "AffectedList": [
+      {
+        "Family": "unix",
+        "Platforms": [
+          "Red Hat Ansible Engine 2.9 for RHEL 8"
+        ]
+      }
+    ],
+    "References": [
+      {
+        "Source": "RHSA",
+        "RefID": "RHSA-2020:0215",
+        "RefURL": "https://access.redhat.com/errata/RHSA-2020:0215"
+      },
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2019-14904",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2019-14904"
+      },
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2019-14905",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2019-14905"
+      }
+    ],
+    "Description": "Ansible is a simple model-driven configuration management, multi-node\ndeployment, and remote-task execution system. Ansible works over SSH and\ndoes not require any software or daemons to be installed on remote nodes.\nExtension modules can be written in any language and are transferred to\nmanaged machines automatically.\n\nThe following packages have been upgraded to a newer upstream version:\nansible (2.9.4)\n\nBug Fix(es):\n* CVE-2019-14904 Ansible: vulnerability in solaris_zone module via crafted\nsolaris zone\n* CVE-2019-14905 Ansible: malicious code could craft filename in\nnxos_file_copy module\n\nSee:\nhttps://github.com/ansible/ansible/blob/v2.9.4/changelogs/CHANGELOG-v2.9.rst\nfor details on bug fixes in this release.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "Copyright 2020 Red Hat, Inc.",
+      "Issued": {
+        "Date": "2020-01-23"
+      },
+      "Updated": {
+        "Date": "2020-01-23"
+      },
+      "Cves": [
+        {
+          "CveID": "CVE-2019-14904",
+          "Cvss2": "",
+          "Cvss3": "7.3/CVSS:3.0/AV:L/AC:L/PR:H/UI:N/S:C/C:H/I:L/A:L",
+          "Cwe": "CWE-20 CWE-78",
+          "Impact": "",
+          "Href": "https://access.redhat.com/security/cve/CVE-2019-14904",
+          "Public": "20191127"
+        },
+        {
+          "CveID": "CVE-2019-14905",
+          "Cvss2": "",
+          "Cvss3": "5.6/CVSS:3.1/AV:L/AC:L/PR:H/UI:N/S:U/C:H/I:L/A:L",
+          "Cwe": "CWE-20 CWE-73",
+          "Impact": "",
+          "Href": "https://access.redhat.com/security/cve/CVE-2019-14905",
+          "Public": "20191127"
+        }
+      ],
+      "Bugzilla": {
+        "Href": "https://bugzilla.redhat.com/1776944",
+        "ID": "1776944"
+      },
+      "AffectedCpeList": {
+        "Cpe": [
+          "cpe:/a:redhat:ansible_engine:2.9",
+          "cpe:/a:redhat:ansible_engine:2.9::el8"
+        ]
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:202010744006"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 8 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:202010744005"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible is earlier than 0:2.9.4-1.el8ae",
+                    "TestRef": "oval:com.redhat.rhsa:tst:20200215001"
+                  },
+                  {
+                    "Comment": "ansible is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ansible-test is earlier than 0:2.9.4-1.el8ae",
+                    "TestRef": "oval:com.redhat.rhsa:tst:20200215003"
+                  },
+                  {
+                    "Comment": "ansible-test is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202010744004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
