+++ oval/redhat2/testdata/golden/dotnet-3.1-including-unpatched/CVE-2020-0606.unaffected.json
@@ -1,0 +1,92 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.unaffected:def:20200606",
+  "Version": "631",
+  "Metadata": {
+    "Title": "Unaffected components for: CVE-2020-0606 dotnet: Malfunctioning StickyNotes annotation XML files malicious execution prevetion (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2020-0606",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2020-0606"
+      }
+    ],
+    "Description": "Red Hat's versions of the associated software have been determined to NOT be affected by CVE-2020-0606.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.rhsa:tst:20200134024"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 7 is installed",
+            "TestRef": "oval:com.redhat.rhsa:tst:20200134023"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "rh-dotnet31 is installed",
+                    "TestRef": "oval:com.redhat.unaffected:tst:20200605001"
+                  },
+                  {
+                    "Comment": "rh-dotnet31 is not installed",
+                    "TestRef": "oval:com.redhat.unaffected:tst:20200605002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "rh-dotnet31-runtime is installed",
+                    "TestRef": "oval:com.redhat.unaffected:tst:20200605003"
+                  },
+                  {
+                    "Comment": "rh-dotnet31-runtime is not installed",
+                    "TestRef": "oval:com.redhat.unaffected:tst:20200605004"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
