+++ oval/redhat2/testdata/golden/dotnet-3.1-including-unpatched/states.json
@@ -1,0 +1,238 @@
+{
+  "RpminfoState": [
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20200134001",
+      "Version": "637",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.1-4.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20200134002",
+      "Version": "637",
+      "Arch": {
+        "Text": "",
+        "Datatype": "",
+        "Operation": ""
+      },
+      "Evr": {
+        "Text": "",
+        "Datatype": "",
+        "Operation": ""
+      },
+      "SignatureKeyID": {
+        "Text": "199e2f91fd431d51",
+        "Operation": "equals"
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20200134003",
+      "Version": "637",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.101-4.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202249001",
+      "Version": "633",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.4-2.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202249003",
+      "Version": "633",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.104-2.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202475001",
+      "Version": "635",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.5-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202475003",
+      "Version": "635",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.105-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202939001",
+      "Version": "636",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.6-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20202939003",
+      "Version": "636",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.106-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20203421001",
+      "Version": "634",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.7-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20203421003",
+      "Version": "634",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.107-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20203697001",
+      "Version": "634",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.8-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    },
+    {
+      "ID": "oval:com.redhat.rhsa:ste:20203697003",
+      "Version": "634",
+      "Arch": {
+        "Text": "x86_64",
+        "Datatype": "string",
+        "Operation": "equals"
+      },
+      "Evr": {
+        "Text": "0:3.1.108-1.el7",
+        "Datatype": "evr_string",
+        "Operation": "less than"
+      },
+      "SignatureKeyID": {
+        "Text": "",
+        "Operation": ""
+      }
+    }
+  ]
+}
