+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/CVE-2014-3209.json
@@ -1,0 +1,120 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:20143209",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2014-3209 ldns: ldns-keygen generates keys with world readable permissions (low)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2014-3209",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2014-3209"
+      }
+    ],
+    "Description": "STATEMENT: This issue affects the versions of ldns as shipped with Red Hat Enterprise Linux 6 and Red Hat Enterprise Linux 7. Red Hat Product Security has rated this issue as having Low security impact. This issue is not currently planned to be addressed in future updates. For additional information, refer to the Issue Severity Classification: https://access.redhat.com/security/updates/classification/.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Low",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ldns-devel is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209001"
+                  },
+                  {
+                    "Comment": "ldns-devel is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ldns-doc is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209003"
+                  },
+                  {
+                    "Comment": "ldns-doc is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209004"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ldns is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209005"
+                  },
+                  {
+                    "Comment": "ldns is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209006"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "ldns-python is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209007"
+                  },
+                  {
+                    "Comment": "ldns-python is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:20143209008"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
