+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/CVE-2016-5361.json
@@ -1,0 +1,65 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:20165361",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2016-5361 IKEv1 protocol is vulnerable to DoS amplification attack (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2016-5361",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2016-5361"
+      }
+    ],
+    "Description": "DOCUMENTATION: A traffic amplification flaw was found in the Internet Key Exchange version 1 (IKEv1) protocol. A remote attacker could use a libreswan server with IKEv1 enabled in a network traffic amplification denial of service attack against other hosts on the network by sending UDP packets with a spoofed source address to that server.\n            STATEMENT: This is a protocol flaw which affects IKEv1. All complaint implementations are therefore affected by this flaw. Red Hat Product Security team, does not consider IKEv2 to be affected. For more details please refer to https://bugzilla.redhat.com/show_bug.cgi?id=1308508#c2",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          },
+          {
+            "Comment": "libreswan is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20165361001"
+          },
+          {
+            "Comment": "libreswan is signed with Red Hat redhatrelease2 key",
+            "TestRef": "oval:com.redhat.cve:tst:20165361002"
+          }
+        ],
+        "Criterias": null
+      }
+    ]
+  }
+}
