+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/CVE-2016-5391.unaffected.json
@@ -1,0 +1,65 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.unaffected:def:20165391",
+  "Version": "631",
+  "Metadata": {
+    "Title": "Unaffected components for: CVE-2016-5391 libreswan: IKEv2 bogus proposal lacking DH transform causes pluto daemon restart (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2016-5391",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2016-5391"
+      }
+    ],
+    "Description": "Red Hat's versions of the associated software have been determined to NOT be affected by CVE-2016-5391.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          },
+          {
+            "Comment": "libreswan is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20165361001"
+          },
+          {
+            "Comment": "libreswan is not installed",
+            "TestRef": "oval:com.redhat.unaffected:tst:20165391002"
+          }
+        ],
+        "Criterias": null
+      }
+    ]
+  }
+}
