+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/CVE-2018-5389.json
@@ -1,0 +1,65 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:20185389",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2018-5389 IKEv1: IKEv1 protocol vulnerability in the authentication mode with pre-shared keys in the main mode of operation (moderate)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2018-5389",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2018-5389"
+      }
+    ],
+    "Description": "DOCUMENTATION: It was found that IKEv1 (and potentially IKEv2) authentication when using a pre-shared key (PSK) is vulnerable to offline dictionary attacks in Main Mode as well as in Aggressive Mode. A man-in-the-middle attacker who intercepted the handshake of two peers authenticating with a PSK, could apply a brute-force attack to recover the shared secret.\n            STATEMENT: PSK based authentication should only be used when the randomness and confidentially of the shared secret can be guaranteed. PSKs should also not be used as Group Secrets, where the security of the PSK is only as strong as the weakest participant in the group. Public Key or EAP authentication methods should be used whenever possible. If PSK must be used, it is essential to ensure the shared secret has a high degree of randomness and is not derived from a password with low entropy, as specified clearly in the IKEv2 specification in RFC 7296.\n\nTo use passwords for authentication of IKE/IPsec peers, the IKEv2 protocol supports various methods that are not based on (inherently weak) PSKs and which are not vulnerable to offline dictionary attacks:\n\nRFC 5998: EAP-Only Authentication in IKEv2\nRFC 6617: Secure Pre-Shared Key (PSK) Authentication for IKE\nRFC 6631: Password Authenticated Connection Establishment with IKEv2\nRFC 6628: Efficient Augmented Password-Only Authentication and Key Exchange for IKEv2\n\nAs implementations supporting IKE assume the security of provided PSKs, and no mechanism within the protocol allows for password-stretching, we do not anticipate any software fixes becoming available.\n\nThe research paper that describes the problems of using weak PSKs also listed another security issue with respect to RSA keys that has different CVE numbers. Libreswan is not vulnerable to those attacks as it requires IKEv1 using either  (\"Encryption with RSA\" (value 5) or \"Revised encryption with RSA\" (value 6). Both of these modes are not implemented by libreswan.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Moderate",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          },
+          {
+            "Comment": "libreswan is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20165361001"
+          },
+          {
+            "Comment": "libreswan is signed with Red Hat redhatrelease2 key",
+            "TestRef": "oval:com.redhat.cve:tst:20165361002"
+          }
+        ],
+        "Criterias": null
+      }
+    ]
+  }
+}
