+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/CVE-2020-28935.json
@@ -1,0 +1,120 @@
+{
+  "Class": "vulnerability",
+  "ID": "oval:com.redhat.cve:def:202028935",
+  "Version": "631",
+  "Metadata": {
+    "Title": "CVE-2020-28935 unbound: symbolic link traversal when writing PID file (low)",
+    "AffectedList": null,
+    "References": [
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2020-28935",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2020-28935"
+      }
+    ],
+    "Description": "DOCUMENTATION: A symbolic link traversal vulnerability was found in the way unbound writes its PID file while starting up. A local attacker with access to the unbound user could set up a link to another file, owned by root, and make unbound overwrite it during its next restart, destroying the original content.\n\n            MITIGATION: If SELinux is enabled in Enforcing mode (the default value in Red Hat Enterprise Linux 8), this kind of attack is prevented as unbound would be blocked from accessing the symbolic link file.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Low",
+      "Rights": "",
+      "Issued": {
+        "Date": ""
+      },
+      "Updated": {
+        "Date": ""
+      },
+      "Cves": null,
+      "Bugzilla": {
+        "Href": "",
+        "ID": ""
+      },
+      "AffectedCpeList": {
+        "Cpe": null
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          }
+        ],
+        "Criterias": [
+          {
+            "Operator": "OR",
+            "Criterions": null,
+            "Criterias": [
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "unbound-libs is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935001"
+                  },
+                  {
+                    "Comment": "unbound-libs is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935002"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "unbound-devel is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935003"
+                  },
+                  {
+                    "Comment": "unbound-devel is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935004"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "unbound-python is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935005"
+                  },
+                  {
+                    "Comment": "unbound-python is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935006"
+                  }
+                ],
+                "Criterias": null
+              },
+              {
+                "Operator": "AND",
+                "Criterions": [
+                  {
+                    "Comment": "unbound is installed",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935007"
+                  },
+                  {
+                    "Comment": "unbound is signed with Red Hat redhatrelease2 key",
+                    "TestRef": "oval:com.redhat.cve:tst:202028935008"
+                  }
+                ],
+                "Criterias": null
+              }
+            ]
+          }
+        ]
+      }
+    ]
+  }
+}
