+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/RHBA-2014-1396.json
@@ -1,0 +1,90 @@
+{
+  "Class": "patch",
+  "ID": "oval:com.redhat.rhba:def:20141396",
+  "Version": "641",
+  "Metadata": {
+    "Title": "RHBA-2014:1396: redhat-upgrade-tool bug fix update (Important)",
+    "AffectedList": [
+      {
+        "Family": "unix",
+        "Platforms": [
+          "Red Hat Enterprise Linux 6 Extras"
+        ]
+      }
+    ],
+    "References": [
+      {
+        "Source": "RHSA",
+        "RefID": "RHBA-2014:1396",
+        "RefURL": "https://access.redhat.com/errata/RHBA-2014:1396"
+      },
+      {
+        "Source": "CVE",
+        "RefID": "CVE-2014-3585",
+        "RefURL": "https://access.redhat.com/security/cve/CVE-2014-3585"
+      }
+    ],
+    "Description": "The Red Hat Upgrade Tool is used for performing an in-place upgrade of your\ncurrent system to the next major version of Red Hat Enterprise Linux. It\ndetermines what packages are needed for the upgrade and gathers them from the\nsource or sources given. It also fetches and sets up the boot images needed to\nrun the upgrade and sets up the system to perform the upgrade on the next system\nboot. Running the Red Hat Upgrade Tool requires running the Preupgrade Assistant\nas a prerequisite.\n\nThis update fixes the following bugs:\n\n* Previously, the Red Hat Subscription Management (RHSM) repository setup did\nnot correctly specify the appropriate URLs. As a consequence,\npreupgrade-assistant postupgrade scripts could not use the Yum utility, and an\nerror message was displayed. With this update, the \"skip_if_unavailable\" option\nhas been set to \"True\" for all generated repository files, and the errors no\nlonger occur. (BZ#1106401)\n\n* Previously, using the \"--addrepo\" option with an incorrect argument returned a\nconfusing error message that no upgrade was available from the repository. As a\nconsequence, it was not clear whether the specified URL was incorrect or there\nwas no upgrade available from the repository. This update improves the error\nmessage which now suggests checking if the specified URL is correct.\n(BZ#1084985)\n\n* After an upgrade, when one of the postupgrade scripts attempted to downgrade\nseveral packages, downgrade failed because of errors downloading the GPG keys.\nThis bug has been fixed by specifying that the postupgrade scripts use the same\nGPG keys as the Red Hat Upgrade Tool. As a result, postupgrade scripts downgrade\nthe selected packages successfully. (BZ#1115532)\n\n* Previously, the Red Hat Upgrade Tool did not perform a check whether the\nrepositories the system is being upgraded to are signed by Red Hat. This update\nadds a check to the Red Hat Upgrade Tool which checks the repository signatures\nand fails the upgrade if the packages are not signed by Red Hat. (BZ#1123915)\n\n* Previously, running the redhat-upgrade-tool command with the \"--clean\" option\ncould fail because the Red Hat Upgrade Tool did not check for the existence of\nthe /var/lib/systemd file. This bug has been fixed and the \"redhat-upgrade-tool\n--clean\" command now executes successfully. (BZ#1128162)\n\n* Previously, after an upgrade, the Red Hat Upgrade Tool did not disable the Yum\nrepositories for the old system. This bug has been fixed and now Red Hat Upgrade\nTool correctly disables all old repositories excluding the\n/etc/yum.repos.d/redhat-upgrade-* repositories. (BZ#1130686)\n\n* Previously, certain packages handled by the Preupgrade Assistant utility and\nconsequently by the Red Hat Upgrade Tool had broken upgrade paths. As a\nconsequence, these packages were not upgraded during system assessment. This\nupdate fixes the upgrade paths and the correct files are now downloaded and\ninstalled by both tools. (BZ#1138615)\n\n* Previously, when performing an upgrade over the network, if there was a\nnetwork timeout error when downloading the .treeinfo file, the Red Hat Upgrade\nTool exited with a traceback error. After this update, if there is a network\ntimeout, the Red Hat Upgrade Tool returns a \"Downloading failed\" error and does\nnot produce a traceback error. (BZ#1076120)\n\nUsers of redhat-upgrade-tool are advised to upgrade to this updated package,\nwhich fixes these bugs.",
+    "Advisory": {
+      "From": "secalert@redhat.com",
+      "Severity": "Important",
+      "Rights": "Copyright 2014 Red Hat, Inc.",
+      "Issued": {
+        "Date": "2014-10-14"
+      },
+      "Updated": {
+        "Date": "2014-10-14"
+      },
+      "Cves": [
+        {
+          "CveID": "CVE-2014-3585",
+          "Cvss2": "7.6/AV:N/AC:H/Au:N/C:C/I:C/A:C",
+          "Cvss3": "",
+          "Cwe": "CWE-347",
+          "Impact": "",
+          "Href": "https://access.redhat.com/security/cve/CVE-2014-3585",
+          "Public": "20141014"
+        }
+      ],
+      "Bugzilla": {
+        "Href": "https://bugzilla.redhat.com/1128162",
+        "ID": "1128162"
+      },
+      "AffectedCpeList": {
+        "Cpe": [
+          "cpe:/a:redhat:rhel_extras_other:6",
+          "cpe:/a:redhat:rhel_extras_other:6"
+        ]
+      }
+    }
+  },
+  "Criteria": {
+    "Operator": "OR",
+    "Criterions": [
+      {
+        "Comment": "Red Hat Enterprise Linux must be installed",
+        "TestRef": "oval:com.redhat.cve:tst:20143209010"
+      }
+    ],
+    "Criterias": [
+      {
+        "Operator": "AND",
+        "Criterions": [
+          {
+            "Comment": "Red Hat Enterprise Linux 6 is installed",
+            "TestRef": "oval:com.redhat.cve:tst:20143209009"
+          },
+          {
+            "Comment": "redhat-upgrade-tool is earlier than 1:0.7.32-1.el6",
+            "TestRef": "oval:com.redhat.rhba:tst:20141396001"
+          },
+          {
+            "Comment": "redhat-upgrade-tool is signed with Red Hat redhatrelease2 key",
+            "TestRef": "oval:com.redhat.rhba:tst:20141396002"
+          }
+        ],
+        "Criterias": null
+      }
+    ]
+  }
+}
