+++ oval/redhat2/testdata/golden/rhel-6-extras-including-unpatched/objects.json
@@ -1,0 +1,59 @@
+{
+  "RpminfoObjects": [
+    {
+      "ID": "oval:com.redhat.cve:obj:20143209001",
+      "Version": "631",
+      "Name": "ldns-devel"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:20143209002",
+      "Version": "631",
+      "Name": "ldns-doc"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:20143209003",
+      "Version": "631",
+      "Name": "ldns"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:20143209004",
+      "Version": "631",
+      "Name": "ldns-python"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:20165361001",
+      "Version": "631",
+      "Name": "libreswan"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:202028935001",
+      "Version": "631",
+      "Name": "unbound-libs"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:202028935002",
+      "Version": "631",
+      "Name": "unbound-devel"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:202028935003",
+      "Version": "631",
+      "Name": "unbound-python"
+    },
+    {
+      "ID": "oval:com.redhat.cve:obj:202028935004",
+      "Version": "631",
+      "Name": "unbound"
+    },
+    {
+      "ID": "oval:com.redhat.rhba:obj:20141396001",
+      "Version": "641",
+      "Name": "redhat-upgrade-tool"
+    },
+    {
+      "ID": "oval:com.redhat.unaffected:obj:20190804001",
+      "Version": "631",
+      "Name": "WALinuxAgent"
+    }
+  ]
+}
