--- oval2/redhat/redhat_test.go
@@ -1,1 +1,0 @@
-package redhat
